+++ frontend/src/components/Header/type.ts
@@ -1,5 +1,5 @@
 import { AvailablePickedColor } from 'styles/type';
 
-export interface WrapperProps {
+export interface HeaderProps {
   bgColor: AvailablePickedColor;
 }
