+++ frontend/src/components/index.ts
@@ -1,7 +1,7 @@
 export * from 'components/@shared/Text';
 export * from 'components/@shared/FlexBox';
 export * from 'components/@shared/Button';
-export * from 'components/AuthInput';
+export * from 'components/Input';
 export * from 'components/LinkText';
 export * from 'components/Logo';
 export * from 'components/ValidationMessage';
@@ -16,11 +16,11 @@ export * from 'components/UnderLineText';
 export * from 'components/ErrorBoundary';
 export * from 'components/ThumbnailWrapper';
 export * from 'components/FixedButton';
-export * from 'components/CertAuth';
-export * from 'components/CertUnAuth';
 export * from 'components/Profile';
 export * from 'components/CardBox';
 export * from 'components/CardGridContainer';
 export * from 'components/EmptyContent';
 export * from 'components/SnackBar';
 export * from 'components/LoadingSpinner';
+export * from 'components/PrivateOutlet';
+export * from 'components/PublicOutlet';
