+++ frontend/src/pages/index.tsx
@@ -6,3 +6,4 @@ export { ProfileEditPage } from 'pages/ProfileEditPage';
 export { ChallengeDetailPage } from 'pages/ChallengeDetailPage';
 export { CertPage } from 'pages/CertPage';
 export { VocPage } from 'pages/VocPage';
+export { LandingPage } from 'pages/LandingPage';
